Add Zen and the Art of Motorcycle Maintenance: An Inquiry Into Values
The book is complicated and requires focus to be understood, bit it deserves that. The author's search for Quality is rather thrilling to observe. The Quality described in the book can challenge lots of people to become more conscious about their work. Overall, great book for those who enjoy challenging reading/