Added "Pilgrim at Tinker Creek"
Really wonderful book about nature, the passage of time, and really paying attention to the world around you. The whole book could almost be categorized as prose poetry; it's not a large volume but took me a while to get through just due to the sheer density of lush and sensitive description. It chronicles a year in which both nothing and everything seems to happen; Dillard mostly hangs out around her home and its backyard creek (or at least, the book is constrained mostly to these experiences) but manages to evoke transcendent beauty and spirituality, capturing cycles of predator and prey, weather, and her own reactions to the unfolding of life in its infinite forms.