Add Bolo'bolo
* Added "Bolo'bolo"

Weird and wonderful little book that proposes a whole societal structure of radical autonomous communities — complete with an invented language to describe the parts of this new community vision, and their machinations. Very eye-opening; in some ways it's a strange and impractical utopian dream; in others it's a sobering and critical plea for reorganizing society along lines of greater sustainability and social cohesion. Very interesting whichever way you look at it!