Added "Stories of Your Life and Others"
"Stories of Your Life and Others" by Ted Chiang is a collection of short science fiction stories. I think it deserves to be on the list because each one has a unique way of opening up your perspective and making you realize that your own worldview is tiny. One of my favorite short stories is "Stories of Your Life" which makes you think about how the language you speak/write can radically shape the way you observe the physical universe (and it was also adapted into the movie, Arrival).